requirements.txt
python-dotenv>=1.0.0
jupyterlab
ipywidgets
requests>=2.31.0
numpy
pandas
scipy
scikit-learn
matplotlib
gensim
torch
transformers
tqdm
openai
gradio
langchain
tiktoken
faiss-cpu
langchain-openai
langchain_experimental
langchain_chroma
langchain[docarray]
datasets
sentencepiece
matplotlib
google-generativeai
anthropic
scikit-learn
unstructured
chromadb
plotly
jupyter-dash
beautifulsoup4>=4.12.0
pydub
modal
ollama>=0.1.0
accelerate
sentencepiece
bitsandbytes
psutil
setuptools
speedtest-cli
sentence_transformers
feedparser
newspaper3k>=0.2.8
selenium>=4.15.0
webdriver-manager>=4.0.0
streamlit>=1.45.0
